# storage/api-docs/api-docs.yaml
...
openapi: 3.0.0
info:
  title: Development Challenge for PHP
  description: This is a sample API documentation for the Development Challenge for
    PHP
  version: 1.0.0
paths:
  "/api/actors":
    get:
      tags:
      - Actors
      summary: Display a listing of the actors
      operationId: 91a2ff1cae626a1a599ae2e37c48053c
      responses:
        '200':
          description: A listing of the actors
          content:
            application/json:
              schema:
                properties:
                  actors:
                    "$ref": "#/components/schemas/Actor"
                type: object
    post:
      tags:
      - Actors
      summary: Store a newly created actor in storage
      operationId: 49a96663e0b82b80d414aadfe3ce9646
      requestBody:
        description: Pass actor data
        required: true
        content:
          application/json:
            schema:
              properties:
                name:
                  type: string
                genre:
                  type: string
                birth_date:
                  type: string
              type: object
      responses:
        '200':
          description: Actor created successfully
          content:
            application/json:
              schema:
                properties:
                  message:
                    type: string
                  actor:
                    "$ref": "#/components/schemas/Actor"
                type: object
        '422':
          description: Invalid data passed
          content:
            application/json:
              schema:
                properties:
                  message:
                    type: string
                  errors:
                    type: object
                type: object
  "/api/actors/{actor}":
    get:
      tags:
      - Actors
      summary: Display the specified actor
      operationId: b6540fce95111401e6d7ee1d2fe2b9ab
      parameters:
      - name: actor
        in: path
        description: Actor ID
        required: true
        schema:
          type: integer
      responses:
        '200':
          description: Actor details
          content:
            application/json:
              schema:
                properties:
                  actor:
                    "$ref": "#/components/schemas/Actor"
                type: object
        '404':
          description: Actor not found
          content:
            application/json:
              schema:
                properties:
                  message:
                    type: string
                type: object
    put:
      tags:
      - Actors
      summary: Update the specified actor in storage
      operationId: 7e65a81bc322d2dd1f04e129e90b971b
      parameters:
      - name: actor
        in: path
        description: Actor ID
        required: true
        schema:
          type: integer
      requestBody:
        description: Pass actor data
        required: true
        content:
          application/json:
            schema:
              properties:
                name:
                  type: string
                genre:
                  type: string
                birth_date:
                  type: string
              type: object
      responses:
        '200':
          description: Actor updated successfully
          content:
            application/json:
              schema:
                properties:
                  message:
                    type: string
                  actor:
                    "$ref": "#/components/schemas/Actor"
                type: object
        '404':
          description: Actor not found
          content:
            application/json:
              schema:
                properties:
                  message:
                    type: string
                type: object
        '422':
          description: Invalid data passed
          content:
            application/json:
              schema:
                properties:
                  message:
                    type: string
                  errors:
                    type: object
                type: object
    delete:
      tags:
      - Actors
      summary: Remove the specified actor from storage
      operationId: 3427d4db2d156404532c1980d5db08c4
      parameters:
      - name: actor
        in: path
        description: Actor ID
        required: true
        schema:
          type: integer
      responses:
        '200':
          description: Actor deleted successfully
          content:
            application/json:
              schema:
                properties:
                  message:
                    type: string
                  actor:
                    "$ref": "#/components/schemas/Actor"
                type: object
        '404':
          description: Actor not found
          content:
            application/json:
              schema:
                properties:
                  message:
                    type: string
                type: object
        '422':
          description: Actor has movies or tv shows
          content:
            application/json:
              schema:
                properties:
                  message:
                    type: string
                type: object
...
components:
  schemas:
    GenderEnum:
      description: 'Gender options: Male, Female'
      type: string
      enum:
      - M
      - F
...
    Actor:
      title: Actor
      description: Actor model
      properties:
        id:
          description: Actor id
          type: integer
          example: '1'
        name:
          description: Actor name
          type: string
          example: John Doe
        birth_date:
          description: Actor birth date
          type: string
          format: date
          example: '1990-01-01'
        gender:
          "$ref": "#/components/schemas/GenderEnum"
      type: object
...